image-producer-fastapi/requirements.txt
fastapi
uvicorn
aiokafka
python-multipart # Required for handling file uploads (UploadFile)
google-cloud-storage